chore: organize in packages
