+++ app/src/main/res/layout/activity_main.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context=".Main">
+    tools:context=".ui.Main">
 
     <androidx.appcompat.widget.Toolbar
         android:id="@+id/header"
