+++ app/src/main/res/layout/activity_splash_aitruck.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".SplashAITruck">
+    tools:context=".ui.SplashAITruck">
 
     <ImageView
         android:id="@+id/video_splash"
